Moved VerticalLayersManager to compas_slicer.geometry.layer
