+++ src/compas_slicer/post_processing/sort_into_vertical_layers.py
@@ -1,4 +1,4 @@
-from compas_slicer.slicers.slice_utilities import VerticalLayersManager
+from compas_slicer.geometry import VerticalLayersManager
 import logging
 
 logger = logging.getLogger('logger')
