+++ src/compas_slicer/slicers/slice_utilities/__init__.py
@@ -5,7 +5,6 @@ from __future__ import print_function
 from .graph_connectivity import *  # noqa: F401 E402 F403
 from .contours_base import *  # noqa: F401 E402 F403
 from .scalar_field_contours import *  # noqa: F401 E402 F403
-from .vertical_layers_manager import *  # noqa: F401 E402 F403
 
 
 __all__ = [name for name in dir() if not name.startswith('_')]
